requirements.txt
fastapi >=0.68.0,<0.69.0
uvicorn[standard]
PyMySQL
python-decouple
selenium
pyodbc
beautifulsoup4
pandas
SQLAlchemy
python-decouple
lxml
alembic
ipython